finish'

+++ Binomial.java
@@ -16,12 +16,12 @@ public class Binomial {
 		if (k == 0 || n == 0) return 1;
 		return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	 }
-	 
+
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		if (k > n) return 0;
         if (k == 0 || n == 0) {return 1;}
-        int[][] memo = new int[n + 1][k + 1];
+        long[][] memo = new long[n + 1][k + 1];
         for (int i = 0; i <= n; i++) {
             for (int j = 0; j <= k; j++) {
                 memo[i][j] = -1;
@@ -31,16 +31,16 @@ public class Binomial {
         return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
-		// Base caseאבל
+		
 		if ((k > n)) {
 		   	memo[n][k] = 0; 
 		   	return 0;
 		}
-		// Another base case
+		
 		if (n == 0 || k == 0) {
 		   	memo[n][k] = 1; 
 		   	return 1;
